PN-15898

+++ PAS_Database/dbo/User Defined Types/VendorRFQRepairOrderPartType.sql
@@ -20,9 +20,9 @@
     [Condition]               VARCHAR (256)   NULL,
     [WorkPerformedId]         BIGINT          NULL,
     [WorkPerformed]           VARCHAR (250)   NULL,
-    [QuantityOrdered]         INT             NULL,
-    [UnitCost]                DECIMAL (20, 2) NULL,
-    [ExtendedCost]            DECIMAL (20, 2) NULL,
+    [QuantityOrdered]         DECIMAL (18, 6) NULL,
+    [UnitCost]                DECIMAL (18, 6) NULL,
+    [ExtendedCost]            DECIMAL (18, 6) NULL,
     [WorkOrderId]             BIGINT          NULL,
     [WorkOrderNo]             VARCHAR (250)   NULL,
     [SubWorkOrderId]          BIGINT          NULL,
@@ -56,6 +56,4 @@
     [TaggedByTypeName]        VARCHAR (250)   NULL,
     [TagDate]                 DATETIME2 (7)   NULL,
     [IsNoQuote]               BIT             NULL);
-
-
